# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeSphere API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface SimpleOk {
  ok: boolean;
}

export interface SignupBody {
  name: string;
  email: string;
  phone: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface OtpChallenge {
  challengeId: string;
  /**
   * Returned in dev mode so users can complete the flow without SMS gateway
   * @nullable
   */
  devOtp: string | null;
}

export interface VerifyOtpBody {
  challengeId: string;
  code: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  phone: string;
  responderStatus: string;
  /** @nullable */
  helperAlias?: string | null;
  createdAt: string;
}

export interface AuthSession {
  token: string;
  user: User;
}

export interface Contact {
  id: number;
  name: string;
  phone: string;
  relation: string;
  createdAt: string;
}

export interface CreateContactBody {
  name: string;
  phone: string;
  relation: string;
}

export interface TriggerSosBody {
  /** button | voice | lock-screen | shake */
  trigger: string;
  lat: number;
  lng: number;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  discreet?: boolean | null;
  /** @nullable */
  message?: string | null;
}

export interface MotionSampleBody {
  /** m/s */
  speed: number;
  /**
   * m/s² peak
   * @nullable
   */
  accel?: number | null;
  /**
   * User pressed SOS multiple times
   * @nullable
   */
  repeated?: boolean | null;
}

export interface RiskUpdate {
  incidentId: number;
  riskScore: number;
  riskLevel: string;
  escalationSeconds: number;
}

export interface Incident {
  id: number;
  status: string;
  trigger: string;
  startLat: number;
  startLng: number;
  alertsSent: number;
  escalated: boolean;
  discreet: boolean;
  riskScore: number;
  riskLevel: string;
  /** @nullable */
  motionMaxSpeed?: number | null;
  /** @nullable */
  triggerCount24h?: number | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  resolvedAt?: string | null;
  createdAt: string;
}

export interface ActiveIncidentResponse {
  incident: Incident | null;
}

export interface EscalationResult {
  incidentId: number;
  nearbyAlerted: number;
  contactsNotified: number;
}

export interface LocationUpdateBody {
  incidentId: number;
  lat: number;
  lng: number;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  speed?: number | null;
}

export interface LocationPoint {
  id: number;
  incidentId: number;
  lat: number;
  lng: number;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  speed?: number | null;
  encrypted: boolean;
  createdAt: string;
}

export interface UploadMediaBody {
  incidentId: number;
  /** audio | video | photo */
  kind: string;
  /** Base64-encoded media payload (encrypted at rest) */
  data: string;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  durationMs?: number | null;
}

export interface MediaItem {
  id: number;
  incidentId: number;
  kind: string;
  /** @nullable */
  url?: string | null;
  sizeBytes: number;
  /** @nullable */
  durationMs?: number | null;
  /** @nullable */
  mimeType?: string | null;
  encrypted: boolean;
  createdAt: string;
}

export interface AlertItem {
  id: number;
  incidentId: number;
  channel: string;
  target: string;
  status: string;
  attempts: number;
  priority: string;
  /** @nullable */
  lastError?: string | null;
  /** @nullable */
  deliveredAt?: string | null;
  createdAt: string;
}

export interface IncidentResponder {
  id: number;
  incidentId: number;
  alias: string;
  status: string;
  distanceKm: number;
  /** @nullable */
  etaMinutes?: number | null;
  createdAt: string;
  /** @nullable */
  respondedAt?: string | null;
}

export interface IncidentDetail {
  incident: Incident;
  alerts: AlertItem[];
  locations: LocationPoint[];
  media: MediaItem[];
  responders: IncidentResponder[];
}

export interface IncidentShareSummary {
  incidentId: number;
  summary: string;
  mapsUrl: string;
}

export interface AvailabilityBody {
  /** available | unavailable */
  status: string;
}

export interface AvailabilityState {
  status: string;
  alias: string;
}

export interface HelperInvitation {
  id: number;
  incidentId: number;
  alias: string;
  status: string;
  distanceKm: number;
  /** @nullable */
  etaMinutes?: number | null;
  createdAt: string;
  /** @nullable */
  respondedAt?: string | null;
  victimAlias: string;
  incidentLat: number;
  incidentLng: number;
  riskLevel: string;
}

export interface AcceptInvitationBody {
  /** @nullable */
  etaMinutes?: number | null;
}

export type IncidentStatsLast7DaysItem = {
  day: string;
  count: number;
};

export interface IncidentStats {
  totalIncidents: number;
  resolvedIncidents: number;
  cancelledIncidents: number;
  activeIncidents: number;
  totalAlerts: number;
  totalContacts: number;
  last7Days: IncidentStatsLast7DaysItem[];
}