# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: SafeSphere API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /responders/availability:
    post:
      operationId: setResponderAvailability
      tags: [responders]
      summary: Toggle whether you are available to help nearby users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityState"
...
  /responders/invitations/{id}/accept:
    post:
      operationId: acceptInvitation
      tags: [responders]
      summary: Accept an SOS invitation as a helper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInvitationBody"
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelperInvitation"
  /responders/invitations/{id}/decline:
    post:
      operationId: declineInvitation
      tags: [responders]
      summary: Decline an SOS invitation as a helper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelperInvitation"
  /location/update:
    post:
      operationId: updateLocation
      tags: [location]
      summary: Push a live location point for an active incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdateBody"
      responses:
        "200":
          description: Stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOk"
  /location/incident/{id}:
    get:
      operationId: getIncidentLocations
      tags: [location]
      summary: Get all location points for an incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Location points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationPoint"
  /media/upload:
    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload audio/video evidence (encrypted at rest)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadMediaBody"
      responses:
        "201":
          description: Media stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
...
  /incidents/{id}:
    get:
      operationId: getIncidentDetail
      tags: [incidents]
      summary: Detailed incident with alerts, locations, media, responders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentDetail"
  /incidents/{id}/share:
    get:
      operationId: getIncidentShareSummary
      tags: [incidents]
      summary: Generate a shareable plain-text evidence summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Share payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentShareSummary"
...
components:
  schemas:
...
    SimpleOk:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    Incident:
      type: object
      required:
        [
          id,
          status,
          trigger,
          startLat,
          startLng,
          alertsSent,
          escalated,
          discreet,
          riskScore,
          riskLevel,
          createdAt,
        ]
      properties:
        id: { type: integer }
        status: { type: string }
        trigger: { type: string }
        startLat: { type: number }
        startLng: { type: number }
        alertsSent: { type: integer }
        escalated: { type: boolean }
        discreet: { type: boolean }
        riskScore: { type: integer }
        riskLevel: { type: string }
        motionMaxSpeed:
          type: ["number", "null"]
        triggerCount24h:
          type: ["integer", "null"]
        message:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        createdAt: { type: string, format: date-time }
...
    LocationUpdateBody:
      type: object
      required: [incidentId, lat, lng]
      properties:
        incidentId: { type: integer }
        lat: { type: number }
        lng: { type: number }
        accuracy:
          type: ["number", "null"]
        speed:
          type: ["number", "null"]
    LocationPoint:
      type: object
      required: [id, incidentId, lat, lng, encrypted, createdAt]
      properties:
        id: { type: integer }
        incidentId: { type: integer }
        lat: { type: number }
        lng: { type: number }
        accuracy:
          type: ["number", "null"]
        speed:
          type: ["number", "null"]
        encrypted: { type: boolean }
        createdAt: { type: string, format: date-time }
    UploadMediaBody:
      type: object
      required: [incidentId, kind, data]
      properties:
        incidentId: { type: integer }
        kind:
          type: string
          description: audio | video | photo
        data:
          type: string
          description: Base64-encoded media payload (encrypted at rest)
        mimeType:
          type: ["string", "null"]
        durationMs:
          type: ["integer", "null"]
    MediaItem:
      type: object
      required: [id, incidentId, kind, sizeBytes, encrypted, createdAt]
      properties:
        id: { type: integer }
        incidentId: { type: integer }
        kind: { type: string }
        url:
          type: ["string", "null"]
        sizeBytes: { type: integer }
        durationMs:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
        encrypted: { type: boolean }
        createdAt: { type: string, format: date-time }
    AlertItem:
      type: object
      required:
        [id, incidentId, channel, target, status, attempts, priority, createdAt]
      properties:
        id: { type: integer }
        incidentId: { type: integer }
        channel: { type: string }
        target: { type: string }
        status: { type: string }
        attempts: { type: integer }
        priority: { type: string }
        lastError:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
          format: date-time
        createdAt: { type: string, format: date-time }
    IncidentResponder:
      type: object
      required:
        [id, incidentId, alias, status, distanceKm, createdAt]
      properties:
        id: { type: integer }
        incidentId: { type: integer }
        alias: { type: string }
        status: { type: string }
        distanceKm: { type: number }
        etaMinutes:
          type: ["integer", "null"]
        createdAt: { type: string, format: date-time }
        respondedAt:
          type: ["string", "null"]
          format: date-time
    IncidentDetail:
      type: object
      required: [incident, alerts, locations, media, responders]
      properties:
        incident:
          $ref: "#/components/schemas/Incident"
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/AlertItem"
        locations:
          type: array
          items:
            $ref: "#/components/schemas/LocationPoint"
        media:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
        responders:
          type: array
          items:
            $ref: "#/components/schemas/IncidentResponder"
    IncidentShareSummary:
      type: object
      required: [incidentId, summary, mapsUrl]
      properties:
        incidentId: { type: integer }
        summary: { type: string }
        mapsUrl: { type: string }
    AvailabilityBody:
      type: object
      required: [status]
      properties:
        status:
          type: string
          description: available | unavailable
    AvailabilityState:
      type: object
      required: [status, alias]
      properties:
        status: { type: string }
        alias: { type: string }
    HelperInvitation:
      type: object
      required:
        [
          id,
          incidentId,
          alias,
          status,
          distanceKm,
          createdAt,
          victimAlias,
          incidentLat,
          incidentLng,
          riskLevel,
        ]
      properties:
        id: { type: integer }
        incidentId: { type: integer }
        alias: { type: string }
        status: { type: string }
        distanceKm: { type: number }
        etaMinutes:
          type: ["integer", "null"]
        createdAt: { type: string, format: date-time }
        respondedAt:
          type: ["string", "null"]
          format: date-time
        victimAlias: { type: string }
        incidentLat: { type: number }
        incidentLng: { type: number }
        riskLevel: { type: string }
    AcceptInvitationBody:
      type: object
      properties:
        etaMinutes:
          type: ["integer", "null"]
...