# lib/api-zod/src/generated/types/alertItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeSphere API specification
 * OpenAPI spec version: 0.2.0
 */

export interface AlertItem {
  id: number;
  incidentId: number;
  channel: string;
  target: string;
  status: string;
  attempts: number;
  priority: string;
  /** @nullable */
  lastError?: string | null;
  /** @nullable */
  deliveredAt?: Date | null;
  createdAt: Date;
}