# lib/api-zod/src/generated/types/availabilityState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeSphere API specification
 * OpenAPI spec version: 0.2.0
 */

export interface AvailabilityState {
  status: string;
  alias: string;
}