# lib/api-zod/src/generated/types/helperInvitation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeSphere API specification
 * OpenAPI spec version: 0.2.0
 */

export interface HelperInvitation {
  id: number;
  incidentId: number;
  alias: string;
  status: string;
  distanceKm: number;
  /** @nullable */
  etaMinutes?: number | null;
  createdAt: Date;
  /** @nullable */
  respondedAt?: Date | null;
  victimAlias: string;
  incidentLat: number;
  incidentLng: number;
  riskLevel: string;
}