# lib/api-zod/src/generated/types/incidentDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeSphere API specification
 * OpenAPI spec version: 0.2.0
 */
import type { AlertItem } from "./alertItem";
import type { Incident } from "./incident";
import type { IncidentResponder } from "./incidentResponder";
import type { LocationPoint } from "./locationPoint";
import type { MediaItem } from "./mediaItem";

export interface IncidentDetail {
  incident: Incident;
  alerts: AlertItem[];
  locations: LocationPoint[];
  media: MediaItem[];
  responders: IncidentResponder[];
}