# lib/api-zod/src/generated/types/locationPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SafeSphere API specification
 * OpenAPI spec version: 0.2.0
 */

export interface LocationPoint {
  id: number;
  incidentId: number;
  lat: number;
  lng: number;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  speed?: number | null;
  encrypted: boolean;
  createdAt: Date;
}